🚚 rename: constants 위치를 common으로 옮기고 가져오기 업데이트 #36
